flatten packages in common domain, rename DomainModel to IdentityModel
