+++ codebase/domain/domain-art/src/main/java/com/jarqprog/artdomain/author/Author.java
@@ -1,7 +1,7 @@
 package com.jarqprog.artdomain.author;
 
 
-import com.jarqprog.commondomain.absmodel.Model;
+import com.jarqprog.commondomain.Model;
 
 public interface Author extends AuthorData, Model {
