+++ codebase/domain/domain-art/src/main/java/com/jarqprog/artdomain/author/AuthorDTO.java
@@ -1,6 +1,6 @@
 package com.jarqprog.artdomain.author;
 
-import com.jarqprog.commondomain.absmodel.DTO;
+import com.jarqprog.commondomain.DTO;
 
 public interface AuthorDTO extends DTO, AuthorData {
 }
