+++ codebase/domain/domain-art/src/main/java/com/jarqprog/artdomain/author/AuthorData.java
@@ -1,7 +1,7 @@
 package com.jarqprog.artdomain.author;
 
 
-import com.jarqprog.commondomain.absmodel.Identity;
+import com.jarqprog.commondomain.Identity;
 
 public interface AuthorData extends Identity {
