+++ codebase/domain/domain-art/src/main/java/com/jarqprog/artdomain/commentary/Commentary.java
@@ -1,7 +1,7 @@
 package com.jarqprog.artdomain.commentary;
 
 
-import com.jarqprog.commondomain.absmodel.Model;
+import com.jarqprog.commondomain.Model;
 import com.jarqprog.artdomain.picture.Picture;
 
 public interface Commentary extends CommentaryData, Model {
