+++ codebase/domain/domain-art/src/main/java/com/jarqprog/artdomain/commentary/CommentaryDTO.java
@@ -1,6 +1,6 @@
 package com.jarqprog.artdomain.commentary;
 
-import com.jarqprog.commondomain.absmodel.DTO;
+import com.jarqprog.commondomain.DTO;
 
 public interface CommentaryDTO extends DTO, CommentaryData {
 }
