+++ codebase/domain/domain-art/src/main/java/com/jarqprog/artdomain/picture/PictureDTO.java
@@ -1,6 +1,6 @@
 package com.jarqprog.artdomain.picture;
 
-import com.jarqprog.commondomain.absmodel.DTO;
+import com.jarqprog.commondomain.DTO;
 
 public interface PictureDTO extends DTO, PictureData {
 }
