+++ codebase/domain/domain-common/src/main/java/com/jarqprog/commondomain/Identity.java
@@ -1,4 +1,4 @@
-package com.jarqprog.commondomain.absmodel;
+package com.jarqprog.commondomain;
 
 public interface Identity {
