+++ codebase/domain/domain-common/src/main/java/module-info.java
@@ -1,5 +1,5 @@
 module domain.common {
-    exports com.jarqprog.commondomain.absmodel;
+    exports com.jarqprog.commondomain;
 
     requires lombok;
     requires org.apache.commons.lang3;
