+++ codebase/domain/domain-person/src/main/java/com/jarqprog/domainperson/contact/Contact.java
@@ -1,7 +1,7 @@
 package com.jarqprog.domainperson.contact;
 
 
-import com.jarqprog.commondomain.absmodel.Model;
+import com.jarqprog.commondomain.Model;
 
 public interface Contact extends ContactData, Model {
