+++ codebase/domain/domain-person/src/main/java/com/jarqprog/domainperson/login/UserLoginDTO.java
@@ -1,6 +1,6 @@
 package com.jarqprog.domainperson.login;
 
-import com.jarqprog.commondomain.absmodel.DTO;
+import com.jarqprog.commondomain.DTO;
 
 public interface UserLoginDTO extends DTO {
