+++ codebase/domain/domain-person/src/main/java/com/jarqprog/domainperson/readuser/ReadUserDTO.java
@@ -1,6 +1,6 @@
 package com.jarqprog.domainperson.readuser;
 
-import com.jarqprog.commondomain.absmodel.DTO;
+import com.jarqprog.commondomain.DTO;
 import com.jarqprog.domainperson.SystemRole;
 import com.jarqprog.domainperson.contact.ContactDTO;
 import com.jarqprog.domainperson.user.UserDTO;
