+++ codebase/domain/domain-person/src/main/java/com/jarqprog/domainperson/roleuser/RoleUserDTO.java
@@ -1,6 +1,6 @@
 package com.jarqprog.domainperson.roleuser;
 
-import com.jarqprog.commondomain.absmodel.DTO;
+import com.jarqprog.commondomain.DTO;
 
 public interface RoleUserDTO extends DTO, RoleUserData {
 }
