+++ codebase/domain/domain-person/src/main/java/com/jarqprog/domainperson/user/User.java
@@ -1,7 +1,7 @@
 package com.jarqprog.domainperson.user;
 
 
-import com.jarqprog.commondomain.absmodel.Model;
+import com.jarqprog.commondomain.Model;
 import com.jarqprog.domainperson.contact.Contact;
 
 public interface User extends UserData, Model {
