+++ codebase/domain/domain-person/src/main/java/com/jarqprog/domainperson/user/UserDTO.java
@@ -1,6 +1,6 @@
 package com.jarqprog.domainperson.user;
 
-import com.jarqprog.commondomain.absmodel.DTO;
+import com.jarqprog.commondomain.DTO;
 
 public interface UserDTO extends DTO, UserData {
 }
